fix build error by contract function name not corresponding

+++ packages/nextjs/app/staker-ui/_components/StakeContractInteraction.tsx
@@ -81,7 +81,7 @@ export const StakeContractInteraction = ({ address }: { address?: string }) => {
           <div className="flex flex-col items-center w-1/2">
             <p className="block text-xl mt-0 mb-1 font-semibold">You Staked</p>
             <span>
-              {myStake ? formatEther(myStake) : 0} {targetNetwork.nativeCurrency.symbol}
+              {myStake ? formatEther(myStake as bigint) : 0} {targetNetwork.nativeCurrency.symbol}
             </span>
           </div>
         </div>
